temp_dashboard/requirements.txt
streamlit==1.44.0
pillow==11.1.0
numpy==1.24.3
pandas==2.2.3
plotly==6.0.1
gitpython==3.1.44
networkx==3.4.2
matplotlib==3.8.3 